pyproject.toml
[project]
name = "crud-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
]
